+++ src/main/java/my/project/repositories/notification/NotificationRepository.java
@@ -7,9 +7,9 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface NotificationRepository extends JpaRepository<Notification, Long> {
+public interface NotificationRepository extends JpaRepository<Notification, Integer> {
 
-    List<Notification> findAllByToUserId(Long userId);
+    List<Notification> findAllByToUserId(Integer userId);
 
-    void deleteAllByToUserId(Long userId);
+    void deleteAllByToUserId(Integer userId);
 }
